+++ momoim-external-api/src/main/java/com/triplem/momoim/api/review/service/ReviewService.java
@@ -3,10 +3,10 @@ package com.triplem.momoim.api.review.service;
 import com.triplem.momoim.core.common.PaginationInformation;
 import com.triplem.momoim.core.domain.gathering.dto.GatheringPreview;
 import com.triplem.momoim.core.domain.gathering.infrastructure.GatheringRepository;
-import com.triplem.momoim.core.domain.review.ReviewRegister;
 import com.triplem.momoim.core.domain.review.dto.ModifyReview;
 import com.triplem.momoim.core.domain.review.dto.MyReview;
 import com.triplem.momoim.core.domain.review.dto.ReviewDetail;
+import com.triplem.momoim.core.domain.review.implement.ReviewRegister;
 import com.triplem.momoim.core.domain.review.infrastructure.ReviewRepository;
 import com.triplem.momoim.core.domain.review.model.Review;
 import com.triplem.momoim.exception.BusinessException;
